requirements.txt
pandas
sklearn
numpy
plotly
dash
statsmodel
